+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Breaking changes
 
 * The traits `RegisterFn` and `RegisterResultFn` are removed.  `Engine::register_fn` and `Engine::register_result_fn` are now implemented directly on `Engine`.
 * `FnPtr::call_dynamic` now takes `&NativeCallContext` instead of consuming it.
+* All `Module::set_fn_XXX` methods are removed, in favor of `Module::set_native_fn`.
 
 
 Version 0.19.14